Refactor github repos to fit inside a general module
